mywebapp/requirements.txt
Django
djangorestframework
pymongo
django-cors-headers
psycopg2
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-cloud-core
google-cloud-firestore
firebase-admin
facebook-sdk
stripe
grpcio
grpcio-tools